app_flask/requirements.txt
flask==2.3.3
numpy==1.25.2
tensorflow-cpu==2.16.1
joblib==1.3.2
werkzeug==2.3.7
scikit-learn==1.3.1
pillow==11.0.0
requests==2.32.3